+++ packages/react-components/react-drawer/src/components/Drawer/renderDrawer.tsx
@@ -2,6 +2,7 @@
 /** @jsxImportSource @fluentui/react-jsx-runtime */
 
 import { assertSlots } from '@fluentui/react-utilities';
+
 import type { DrawerState, DrawerSlots } from './Drawer.types';
 
 /**
