+++ packages/react-components/react-drawer/src/components/DrawerBody/renderDrawerBody.tsx
@@ -1,7 +1,8 @@
 /** @jsxRuntime automatic */
 /** @jsxImportSource @fluentui/react-jsx-runtime */
 import { assertSlots } from '@fluentui/react-utilities';
-import type { DrawerBodyState, DrawerBodySlots } from './DrawerBody.types';
+
+import type { DrawerBodySlots, DrawerBodyState } from './DrawerBody.types';
 
 /**
  * Render the final JSX of DrawerBody
