+++ packages/react-components/react-drawer/src/components/DrawerBody/useDrawerBody.ts
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import { getNativeElementProps, slot } from '@fluentui/react-utilities';
+
 import type { DrawerBodyProps, DrawerBodyState } from './DrawerBody.types';
 
 /**
