+++ packages/react-components/react-drawer/src/components/DrawerFooter/renderDrawerFooter.tsx
@@ -1,7 +1,8 @@
 /** @jsxRuntime automatic */
 /** @jsxImportSource @fluentui/react-jsx-runtime */
 import { assertSlots } from '@fluentui/react-utilities';
-import type { DrawerFooterState, DrawerFooterSlots } from './DrawerFooter.types';
+
+import type { DrawerFooterSlots, DrawerFooterState } from './DrawerFooter.types';
 
 /**
  * Render the final JSX of DrawerFooter
