+++ packages/react-components/react-drawer/src/components/DrawerHeader/renderDrawerHeader.tsx
@@ -1,6 +1,7 @@
 /** @jsxRuntime automatic */
 /** @jsxImportSource @fluentui/react-jsx-runtime */
 import { assertSlots } from '@fluentui/react-utilities';
+
 import type { DrawerHeaderState, DrawerHeaderSlots } from './DrawerHeader.types';
 
 /**
