+++ packages/react-components/react-drawer/src/components/DrawerHeader/useDrawerHeader.ts
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import { getNativeElementProps, slot } from '@fluentui/react-utilities';
+
 import type { DrawerHeaderProps, DrawerHeaderState } from './DrawerHeader.types';
 
 /**
