+++ packages/react-components/react-drawer/src/components/DrawerInline/DrawerInline.types.ts
@@ -1,5 +1,6 @@
 import type { ComponentProps, ComponentState, Slot } from '@fluentui/react-utilities';
-import { DrawerBaseProps, DrawerBaseState } from '../../util/DrawerBase.types';
+
+import { DrawerBaseProps, DrawerBaseState } from '../../shared/DrawerBase.types';
 
 export type DrawerInlineSlots = {
   root: Slot<'div'>;
