+++ packages/react-components/react-drawer/src/components/DrawerInline/renderDrawerInline.tsx
@@ -1,6 +1,7 @@
 /** @jsxRuntime automatic */
 /** @jsxImportSource @fluentui/react-jsx-runtime */
 import { assertSlots } from '@fluentui/react-utilities';
+
 import type { DrawerInlineState, DrawerInlineSlots } from './DrawerInline.types';
 
 /**
