+++ packages/react-components/react-drawer/src/components/DrawerInline/useDrawerInlineStyles.styles.ts
@@ -4,7 +4,7 @@ import type { SlotClassNames } from '@fluentui/react-utilities';
 import { tokens } from '@fluentui/react-theme';
 
 import type { DrawerInlineSlots, DrawerInlineState } from './DrawerInline.types';
-import { drawerCSSVars, useDrawerBaseClassNames } from '../../util/useDrawerBaseStyles.styles';
+import { drawerCSSVars, useDrawerBaseClassNames } from '../../shared/useDrawerBaseStyles.styles';
 
 export const drawerInlineClassNames: SlotClassNames<DrawerInlineSlots> = {
   root: 'fui-DrawerInline',
