+++ packages/react-components/react-drawer/src/components/DrawerOverlay/renderDrawerOverlay.tsx
@@ -1,8 +1,9 @@
 /** @jsxRuntime automatic */
 /** @jsxImportSource @fluentui/react-jsx-runtime */
 import { assertSlots } from '@fluentui/react-utilities';
-import type { DrawerOverlayState, DrawerOverlaySlots } from './DrawerOverlay.types';
 import { Dialog } from '@fluentui/react-dialog';
+
+import type { DrawerOverlayState, DrawerOverlaySlots } from './DrawerOverlay.types';
 
 /**
  * Render the final JSX of DrawerOverlay
