+++ packages/react-components/react-drawer/src/shared/DrawerBase.types.ts
